Move building share menu to share-menu.js
